linux-mainline: changes rebased onto 4.19 branch
Since the begining of this distro, we kept our modifications
to kernel atop a next-20180725 tree from linux-next. It wasn't
until Vladimir reported that disabling CFQ io scheduler renders
the system incapable of booting init that I considered
a possibility that there could be bugs in the codebase.

Fortunately, our patchset turned out to be easily rebased onto
the 4.19 branch in the linux-stable repo. Coincidentally, this
is also an LTS version and as such it is still being maintained
as of commiting this change.