systemd: implement proper shutdown
To achieve a proper shutdown (power off everything) a command
must be send to the motion controller. This was implemented
only in the sl1fw python code, which is rather difficult to
access from a command-line.

Intuitively, such step should be implemented the closest to
actual shutting down of CPU cores, so that all possible ways
of shutting the printer down would get covered. Thus, the best
place for such code is in ATF's `sunxi_power_down` function.
In order to avoid excessive complexity with setting up the UART
though, systemd with its access to libc was chosen as a good
compromise.