Disable debian-style autonaming of library packages for mali-blobs
Since a dummy libffi solib started to be packaged together with libMali.so to
satisfy its arbitrary NEEDED requirements, the debian autonamer has kicked in.

Because libMali.so.6 has no SONAME of its own, the only one that comes into
consideration is the libffi.so.6. This then however creates libffi{,-dev,-dbg}
packages which are in conflict with the legit libffi package which is depended
upon by other parts of the system. DEBIAN_NOAUTONAME prevents the renaming of
mali-blobs to libffi which resolves the situation.