linux-mainline: drop custom code for reading MAC from eFUSEs
Linux already does it. This feature has been however suppressed by u-boot
filling a random mac address into the device-tree by itself if it finds
`ethernet[0-9]` alias in the DT.

Since the ethernet0 alias had no effect except for this one, it's been
dropped along with other useless aliases. `serial0` was kept as it is
referenced in /chosen/stdout-path property.