udev-usbmount: don't log failed unit resets
Just in case there are failed units lingering from previous mount
a reset is performed to flush them out. Instead of checking whether
there are actually any units to reset, the operation is performed
unconditionally and its result is ignored.

An errorneous result was however still logged (as INFO) which proved
needlessly confusing for readers of that log.