u-boot: fix invalid root= device specification
Turns out that the `root=` parameter was actually never involved
in determining the appropriate fs to mount when booting from the eMMC,
as rauc.slot value was used for that purpose by the initrd script
instead. With the initramfs gone, the value of `root=` has suddenly
began to matter.

rootfs.0 and rootfs.1 are names assigned to partitions on the level
of the GUID Partition Table and those are denoted using the PARTLABEL=
attribute. LABEL= instead refers to a label stored within a filesystem.

```
PATH              LABEL   PARTLABEL
/dev/mmcblk0
/dev/mmcblk0p1    root    root
/dev/mmcblk2
/dev/mmcblk2p1            environment
/dev/mmcblk2p2    root    rootfs.0
/dev/mmcblk2p3    root    rootfs.1
/dev/mmcblk2p4    etc     etc.0
/dev/mmcblk2p5    etc     etc.1
/dev/mmcblk2p6    var     var
/dev/mmcblk2p7    factory factory
/dev/mmcblk2boot0
/dev/mmcblk2boot1
```

Additionaly, while Linux allows for the rootfs to be specified with
PARTLABEL, it doesn't allow the same for LABEL (as the way it's stored
varies with the filesystem type and thus requires the udev userspace
approach.)