bump submodules to their dunfell HEADs
(no pun intended)

In July, a new syntax rule was introduced to bitbake in an effort to put an end
to the long-standing problem with override directives blending visually too much
with the actual names of variables or functions, making the code unnecessarily
more difficult to parse/navigate. The rule was that ':' character became allowed
allowed in variable/function names and recognized as a prefix to overrides.

This has quickly caught on and as the underlying change to the parser was rather
trivial, it's been since backported to older branches.