drop mali-blobs and associated gl-mali/lima override
In preparation for mass-replacement of '_' with ':' in override directives,
Yocto release notes advise to identify all overrides specific to our layer.

Given that we've been having systems with the lima driver out in the wild
for quite some time already without any problems, now seems to be a good
time to finally get rid of the proprietary Mali driver and all the baggage
tied to it.