update to Yocto 3.4 aka honistar (broken)
Layers were switched to honistar, bitbake updated to 1.52.
As we're moving across several version of Yocto, a lot of stuff got broken.
Many of the fixes required to get it back going are already included in this
commit but only those that are trivial enough. The rest will get addressed
across several following commits.